cas4.2 migration - unfinished yet
There is a new feature to encrypt the webflow, unfortunately JCEKS
keystore types are unavailable by default in Java7 and require extra
packages from Oracle.

Need to see how to deactivate the webflow encryption in cas-servlet.xml,
or add a NullTranscoder, which might not be the clever way to deal with
this issue.