mfapp - PoC using the jib-maven-plugin
This is a try of the jib maven plugin. Jib a another way of generating
docker images, the main benefit here is IIRC you do not need to have a
local docker daemon to generate and push images (which can be pretty
convenient in some OpenShift environments).

TODO: for now on, it takes a distroless/jetty image, which is pretty
minimalistic, do not expect sh or bash to live-debug the image.